Extract UiControlsWithNames
