+++ src/app/modules/layer-messier/components/selector-messier-names/selector-messier-names.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { LayerAware } from '#core/models/layers/layer-aware';
 import { TextsVisibilityManagerService } from '#core/services/visibility/texts-visibility-manager.service';
+import { UiControlsWithNames } from '#core/components/ui-controls-with-names';
 
 
 /**
@@ -11,26 +11,10 @@ import { TextsVisibilityManagerService } from '#core/services/visibility/texts-v
   selector: 'ac-controls-select-messier-names',
   templateUrl: './selector-messier-names.component.html'
 })
-export class SelectorMessierNamesComponent extends LayerAware {
+export class SelectorMessierNamesComponent extends UiControlsWithNames {
 
-  private _namesShown: boolean;
-
-  constructor(private readonly _textsVisibilityManager: TextsVisibilityManagerService) {
-    super();
-    this._namesShown = true;
-  }
-
-  public get namesShown(): boolean {
-    return this._namesShown;
-  }
-
-  public set namesShown(show: boolean) {
-    this._namesShown = show;
-    if (this._namesShown) {
-      this._textsVisibilityManager.showTexts(this.layer.code);
-    } else {
-      this._textsVisibilityManager.hideTexts(this.layer.code);
-    }
+  constructor(textsVisibilityManager: TextsVisibilityManagerService) {
+    super(textsVisibilityManager);
   }
 
 }
